add a conversion method from the alias component
We fix an implementation bug on priority (initial backport from alias
missed an ordering issue).

We need to subclass alias to sail smoothly.

However the purpose here is to facilitate the transition from alias to
formulas, and in a couple of releases, aliases will be removed
entirely.

A `tsh convert-aliases` command is provided. It will convert everything
at once.