evaluator: use the auto registry to decide when to make a thread execution
We switch from a brutal thread-spawning strategy to a more controlled
thread-pool based concurrency.

The downside is potential deadlocks (not easy to showcase in a test)
whenever top-level "series" call would be waiting for their leaves to
be evaluated but the pool would be exhausted.

Hypothetical scenario:

 (add (series "A") ... (series "Z"))

... with these series expanding into more series, it is possible that
the pool thread becomes exhausted at this level and does not allow
further execution of the lower parts of the tree, thus blocking any
progress of the waiting threads.

To avoid such a situation, running the formula fully expanded will be
addressed in the next commmit. Doing so garantees that async calls
only happens in the leaves.

We start to immediately take care of the AUTO structure to make a more
robust implementaion of the expansion helper.